Put render code in separate module + refactor
